Refactor growth stage and daily food usage log DTOs to support nullable fields for dead quantity and weights; update related logic in views and view models to handle null values appropriately. Enhance UI responsiveness by dismissing keyboard on tap outside input fields in various widgets. Adjust calculations for recommended weights to ensure correct unit conversions and rounding. Clean up commented-out code in task detail view for better readability.

+++ lib/src/model/dto/task/daily_food_usage_log/daily_food_usage_log_dto.dart
@@ -6,8 +6,8 @@ part 'daily_food_usage_log_dto.g.dart';
 @freezed
 class DailyFoodUsageLogDto with _$DailyFoodUsageLogDto {
   const factory DailyFoodUsageLogDto({
-    required int recommendedWeight,
-    required int actualWeight,
+    required double recommendedWeight,
+    required double actualWeight,
     required String notes,
     required DateTime logTime,
     required String photo,
